Возможность вести контроль транспорта

+++ RjProduction/Model/DocElement/FixCub.cs
@@ -6,10 +6,10 @@ namespace RjProduction.Model.DocElement
     {
         public double Price { get; set; }
         public double Plan { get; set; } = 0;
-        public double Fact { get; set; }
+        public float Fact { get; set; }
 
         public decimal Amount => (decimal)(Price * Fact);
-        public double CubatureAll => Fact;
+        public float CubatureAll => Fact;
 
         public override string ToString() =>"Фактически: " + CubatureAll + " м/куб = " + Amount + " руб.";
     }
